requirements.txt
streamlit
pandas
google-auth
gspread 